+++ sdk/resourcemanager/compute/armcompute/models.go
@@ -4858,6 +4858,13 @@ type SSHConfiguration struct {
 	PublicKeys []*SSHPublicKey
 }
 
+// SSHGenerateKeyPairInputParameters - Parameters for GenerateSshKeyPair.
+type SSHGenerateKeyPairInputParameters struct {
+	// The encryption type of the SSH keys to be generated. See SshEncryptionTypes for possible set of values. If not provided,
+	// will default to RSA
+	EncryptionType *SSHEncryptionTypes
+}
+
 // SSHPublicKey - Contains information about SSH certificate public key and the path on the Linux VM where the public key
 // is placed.
 type SSHPublicKey struct {